+++ fixedformat4j/src/main/java/com/ancientprogramming/fixedformat4j/format/impl/DoubleFormatter.java
@@ -27,7 +27,6 @@ public class DoubleFormatter extends AbstractDecimalFormatter<Double> {
 
   /** {@inheritDoc} */
   public Double asObject(String string, FormatInstructions instructions) {
-    String toConvert = getStringToConvert(string, instructions);
-    return Double.parseDouble("".equals(toConvert) ? "0" : toConvert);
+    return Double.parseDouble(resolveDecimalString(string, instructions));
   }
 }